+++ app/build.gradle
@@ -36,7 +36,7 @@ android {
     }
 
     buildFeatures {
-        viewBinding = true
+        dataBinding true
     }
 }
