+++ app/src/main/java/com/egorpoprotskiy/composition/presentation/GameFinishedFragment.kt
@@ -1,16 +1,12 @@
-package com.egorpoprotskiy.composition.Presentation
+package com.egorpoprotskiy.composition.presentation
 
 import android.os.Bundle
 import android.view.LayoutInflater
 import android.view.View
 import android.view.ViewGroup
-import androidx.activity.OnBackPressedCallback
 import androidx.fragment.app.Fragment
-import androidx.fragment.app.FragmentManager
 import androidx.navigation.fragment.findNavController
 import androidx.navigation.fragment.navArgs
-import com.egorpoprotskiy.composition.Domain.entity.GameResult
-import com.egorpoprotskiy.composition.R
 import com.egorpoprotskiy.composition.databinding.FragmentGameFinishedBinding
 
 class GameFinishedFragment : Fragment() {
@@ -46,7 +42,8 @@ class GameFinishedFragment : Fragment() {
         //9.1 Вызов перезапуска игры
         setupOnClickListeners()
         //9.2 Вызов метода установки значений
-        bindViews()
+//        bindViews()
+        binding.gameResult = args.gameResult
     }
     //9.1 Слушатель кликов на перезапуск игры
     private fun setupOnClickListeners() {
@@ -56,37 +53,39 @@ class GameFinishedFragment : Fragment() {
         }
     }
     //9.2 Создать метод, в котором устанавливаем значения
-    private fun bindViews() {
-        with(binding) {
-            //установка картинки
-            emojiResult.setImageResource(getSmileResId())
-            //установка текстов
-            tvRequiredAnswers.text = String.format(getString(R.string.required_score),
-                //12.4 Получение аргументов из GameFragment
-                args.gameResult.gameSettings.minCountOfRightAnswers)
-            tvScoreAnswers.text = String.format(getString(R.string.score_answer),
-                args.gameResult.countOfRightAnswers)
-            tvRequiredPercentage.text = String.format(getString(R.string.reauired_percentage),
-                args.gameResult.gameSettings.minPercentOfRightAnswers)
-            tvScorePercentage.text = String.format(getString(R.string.score_percentage), getPercentOfRightAnswers())
-        }
-    }
+//    private fun bindViews() {
+        //13.5 Получение аргументов в одну переменную
+//        binding.gameResult = args.gameResult
+//        with(binding) {
+//            //установка картинки
+//            emojiResult.setImageResource(getSmileResId())
+//            //установка текстов
+//            tvRequiredAnswers.text = String.format(getString(R.string.required_score),
+//                //12.4 Получение аргументов из GameFragment
+//                args.gameResult.gameSettings.minCountOfRightAnswers)
+//            tvScoreAnswers.text = String.format(getString(R.string.score_answer),
+//                args.gameResult.countOfRightAnswers)
+//            tvRequiredPercentage.text = String.format(getString(R.string.reauired_percentage),
+//                args.gameResult.gameSettings.minPercentOfRightAnswers)
+//            tvScorePercentage.text = String.format(getString(R.string.score_percentage), getPercentOfRightAnswers())
+//        }
+//    }
     //9.3 Выбор смайлика, грустный или довольный
-    private fun getSmileResId(): Int {
-        return if (args.gameResult.winner) {
-            R.drawable.smile
-        } else {
-            R.drawable.smile_sad
-        }
-    }
-    //9.4 Метод, возвращающий процент правильных ответов
-    private fun getPercentOfRightAnswers() = with(args.gameResult) {
-        if (countOfQuestions == 0) {
-            0
-        } else {
-            ((countOfRightAnswers / countOfQuestions.toDouble()) * 100).toInt()
-        }
-    }
+//    private fun getSmileResId(): Int {
+//        return if (args.gameResult.winner) {
+//            R.drawable.smile
+//        } else {
+//            R.drawable.smile_sad
+//        }
+//    }
+//    //9.4 Метод, возвращающий процент правильных ответов
+//    private fun getPercentOfRightAnswers() = with(args.gameResult) {
+//        if (countOfQuestions == 0) {
+//            0
+//        } else {
+//            ((countOfRightAnswers / countOfQuestions.toDouble()) * 100).toInt()
+//        }
+//    }
     //5.3 Получение аргументов
 //    private fun parseArgs() {
 //        requireArguments().getParcelable<GameResult>(KEY_GAME_RESULT)?.let {
