+++ app/src/main/res/navigation/main_navigation.xml
@@ -7,7 +7,7 @@
 
     <fragment
         android:id="@+id/welcomeFragment"
-        android:name="com.egorpoprotskiy.composition.Presentation.WelcomeFragment"
+        android:name="com.egorpoprotskiy.composition.presentation.WelcomeFragment"
         android:label="WelcomeFragment" >
         <action
             android:id="@+id/action_welcomeFragment_to_chooseLevelFragment"
@@ -15,7 +15,7 @@
     </fragment>
     <fragment
         android:id="@+id/chooseLevelFragment"
-        android:name="com.egorpoprotskiy.composition.Presentation.ChooseLevelFragment"
+        android:name="com.egorpoprotskiy.composition.presentation.ChooseLevelFragment"
         android:label="fragment_choose_level"
         tools:layout="@layout/fragment_choose_level" >
         <action
@@ -24,7 +24,7 @@
     </fragment>
     <fragment
         android:id="@+id/gameFragment"
-        android:name="com.egorpoprotskiy.composition.Presentation.GameFragment"
+        android:name="com.egorpoprotskiy.composition.presentation.GameFragment"
         android:label="fragment_game"
         tools:layout="@layout/fragment_game" >
         <action
@@ -34,15 +34,15 @@
             app:popUpToInclusive="true" />
         <argument
             android:name="level"
-            app:argType="com.egorpoprotskiy.composition.Domain.entity.Level" />
+            app:argType="com.egorpoprotskiy.composition.domain.entity.Level" />
     </fragment>
     <fragment
         android:id="@+id/gameFinishedFragment"
-        android:name="com.egorpoprotskiy.composition.Presentation.GameFinishedFragment"
+        android:name="com.egorpoprotskiy.composition.presentation.GameFinishedFragment"
         android:label="fragment_game_finished"
         tools:layout="@layout/fragment_game_finished" >
         <argument
             android:name="gameResult"
-            app:argType="com.egorpoprotskiy.composition.Domain.entity.GameResult" />
+            app:argType="com.egorpoprotskiy.composition.domain.entity.GameResult" />
     </fragment>
 </navigation>
